requirements.txt
torch
torchvision
numpy
PyYAML
matplotlib
pandas
scikit-learn
tqdm
